requirements.txt
requests
playwright
eventlet
schedule
cf_clearance
biliup
xhs
qrcode
loguru
apscheduler
f2
pywin32
ruamel.yaml
opencv-python
ffmpeg-python
opencc-python-reimplemented
numpy
pysrt
openai-whisper
pydub
Pillow
watchdog
pyyaml
psutil
openai
dashscope